% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate arithmetic mean}
\usage{
calc_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove NA values. Default is TRUE}
}
\value{
A single numeric value representing the mean
}
\description{
Calculate arithmetic mean
}
\examples{
calc_mean(c(1, 2, 3, 4, 5))
}
