% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_median}
\alias{calc_median}
\title{Calculate median}
\usage{
calc_median(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove NA values. Default is TRUE}
}
\value{
A single numeric value representing the median
}
\description{
Calculate median
}
\examples{
calc_median(c(1, 2, 3, 4, 5))
}
