% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate mode}
\usage{
calc_mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove NA values. Default is TRUE}
}
\value{
A numeric vector of the most frequent value(s)
}
\description{
Calculate mode
}
\examples{
calc_mode(c(1, 2, 2, 3, 4))
}
