% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_q1}
\alias{calc_q1}
\title{Calculate first quartile (Q1)}
\usage{
calc_q1(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove NA values. Default is TRUE}
}
\value{
A single numeric value representing Q1
}
\description{
Calculate first quartile (Q1)
}
\examples{
calc_q1(c(1, 2, 3, 4, 5))
}
