% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCorrelations.R
\name{CalculateCorrelations}
\alias{CalculateCorrelations}
\title{Calculates Spearman's correlation}
\usage{
CalculateCorrelations(data)
}
\arguments{
\item{data}{A matrix of exposure concentrations}
}
\value{
A matrix of correlation coefficients
}
\description{
Calculates Spearman's correlation
}
