% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateGeomean.R
\name{CalculateGeomean}
\alias{CalculateGeomean}
\title{Calculates geometrical mean}
\usage{
CalculateGeomean(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A scalar of geometrical mean
}
\description{
Calculates geometrical mean
}
