% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{desc}
\alias{desc}
\title{Creates histogram, box plot and computes five number summary statistics}
\usage{
desc(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Creates histogram, box plot and computes five number summary statistics
}
