% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveStats-package.R
\docType{package}
\name{descriptiveStats-package}
\alias{descriptiveStats}
\alias{descriptiveStats-package}
\title{descriptiveStats: Descriptive Statistics Functions}
\description{
A package providing descriptive statistics functions including mean, median, mode, quartiles and IQR.
}
\author{
\strong{Maintainer}: Raheem Malik \email{r9him.malik@gmail.com}

}
\keyword{internal}
