% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureCharacteristics.R
\name{ExposureCharacteristics}
\alias{ExposureCharacteristics}
\title{Calculates exposure characteristics}
\usage{
ExposureCharacteristics(input_data_descr, input_data_conc)
}
\arguments{
\item{input_data_descr}{Data frame with variables containing exposure characteristics (descriptive about LOD)}

\item{input_data_conc}{Data frame with variables containing exposure concentrations}
}
\value{
A data frame with exposure descriptive statistics
}
\description{
Calculates exposure characteristics
}
