% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_na.R
\name{handle_na}
\alias{handle_na}
\title{Handle NA}
\usage{
handle_na(x, na.rm)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Handles missing values by setting them to default to FALSE.}
}
\value{
The original input, but with empty NA values removed.
}
\description{
A helper function that handles missing values in the input vector by removing NA values.
}
\details{
This function is a helper function that handles the management of NA values in the input.
}
