% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationCharacteristics.R
\name{PopulationCharacteristics}
\alias{PopulationCharacteristics}
\title{Calculates population characteristics}
\usage{
PopulationCharacteristics(input_data, variable_list, path)
}
\arguments{
\item{input_data}{Data frame with variables containing population characteristics}

\item{variable_list}{A character vector with names of population characteristics variables}

\item{path}{A string defining the path to save the file}
}
\value{
A dataframe with population characteristics
}
\description{
Calculates population characteristics
}
