% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric.R
\name{validate_numeric}
\alias{validate_numeric}
\title{Validate Numeric Input}
\usage{
validate_numeric(x)
}
\arguments{
\item{x}{Input vector to check.}
}
\value{
Returns TRUE if validation passes.
}
\description{
Internal helper function to validate whether an input is numeric or not.
}
\keyword{internal}
