% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iqr.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate Interquartile Range (IQR)}
\usage{
calc_iqr(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Interquartile range (Q3 - Q1).
}
\description{
Calculate Interquartile Range (IQR)
}
