% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate interquartile range (IQR)}
\usage{
calc_iqr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Numeric value representing IQR.
}
\description{
Calculate interquartile range (IQR)
}
