% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iqr.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate Interquartile (IQR) Range}
\usage{
calc_iqr(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Handles missing values by setting them to default to FALSE.}
}
\value{
A single numeric value representing the interquartile range of the input.
Returns NA if input is empty after NA removal.
}
\description{
Computes the interquartile range of a numerical vector
}
\details{
This function re-uses calc_q3 and calc_q1 in order to calculate the two required quartiles.
The interquartile range (IQR) is defined as the difference between the first and third
quartiles.
}
\examples{
calc_iqr(c(1, 2, 2, 3))
calc_iqr(c(1, NA, 3), na.rm = TRUE)

}
