% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate the interquartile range of a vector}
\usage{
calc_iqr(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The interquartile range of the input vector. Will return error messages if the vector is not numeric or empty
}
\description{
Calculate the interquartile range of a vector
}
\examples{
calc_iqr(c(1, 2, 3))       # Returns 1
}
