% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate Interquartile Range}
\usage{
calc_iqr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the Interquartile Range.
}
\description{
Computes the Interquartile Range of a numeric vector.
Missing values (NA) are removed before calculation.
}
\examples{
calc_iqr(c(1, 2, 3, 4))
calc_iqr(c(1, 2, NA, 4))

}
