% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate Interquartile Range (IQR)}
\usage{
calc_iqr(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A scalar numeric value representing IQR
}
\description{
Computes the interquartile range (Q3 - Q1).
}
\examples{
# Basic usage
calc_iqr(c(1,2,3))
calc_iqr(c(1,2,3,4))

# NA values are removed with explicit information
calc_iqr(c(1,2,3, NA))

# All values missing -> error
\dontrun{
calc_iqr(c(NA, NA, NA))
}

# Non-numeric input -> error
\dontrun{
calc_iqr("a")
}
}
