% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value, the mean of \code{x}.
}
\description{
Calculates the mean of a numeric vector, ignoring NA values.
}
\examples{
calc_mean(c(1, 2, 3, 4, 5))
# Returns 3
}
