% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the arithmetic mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the arithmetic mean.
}
\description{
Calculates the arithmetic mean of a numeric vector.
}
\examples{
data <- c(1, 2, 3, 4, 5)
calc_mean(data)

}
