% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate arithmetic mean}
\usage{
calc_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector. Must not be empty.}

\item{na.rm}{Logical. Should missing values be removed before calculation?
Defaults to \code{FALSE}.}
}
\value{
A single numeric value representing the arithmetic mean of \code{x}.
}
\description{
Returns arithmetic mean of a numerical vector, with control over
how \code{NA} values are handled.
}
\examples{
calc_mean(c(1, 2, 2, 3, 4, 5, 5, 5, 6, 10))  # 4.3
calc_mean(c(1, 2, NA, 4), na.rm = TRUE) # 2.333333
}
\author{
Cara Andrews
}
