% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the arithmetic mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the arithmetic mean.
}
\description{
Calculates the arithmetic mean of a numeric vector.
Entering Missing values (\code{NA}) are ignored.
}
\details{
\code{NA} values are removed before calculating the mean. If \code{x} is empty
or contains only \code{NA} values, the function returns an informative error.
A vector containing a single non-missing value returns that value.
}
\examples{
calc_mean(c(8, 2, 9, 4, 5))
calc_mean(c(50, 2, NA, 30))
calc_mean(78)

}
