% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_1.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate arithmetic mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric value corresponding to the mean.
If x contains missing values, the function will remove them and compute mean for available values.
If x is null or contains a single value, the function will return an error message.
}
\description{
This function calculates the average or mean of a numerical object
}
\examples{
df <- data.frame(
   patid = c("PO1", "P02", "P03", "P04", "P05"),
   glucose = c(5.6, 7.5, 4.3, 8.4, 10.2)
   )

 calc_mean(df$glucose)

}
