% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the Mean}
\usage{
calc_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric input/vector.}

\item{na.rm}{Handles missing values by setting them to default to FALSE.}
}
\value{
A single numeric value representing the mean of the input vector.
}
\description{
Computes the mean of a numeric vector.
}
\details{
This function is a wrapper around Base R's mean function, with additional
input validation for increased robustness.
}
\examples{
calc_mean(c(1, 2, 3, 4))
calc_mean(c(1, NA, 3), na.rm = TRUE)

}
