% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical. Should missing values be removed? Default is TRUE.}
}
\value{
The arithmetic mean as a numeric value, or NA if the vector is empty
or contains only NA values.
}
\description{
Computes the arithmetic mean of a numeric vector.
}
\examples{
calc_mean(c(1, 2, 3, 4, 5))
calc_mean(c(1, 2, NA, 4, 5), na.rm = TRUE)
calc_mean(c(10, 20, 30))

}
