% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the mean of a vector}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean of the input vector. Will return error messages if the vector is not numeric or empty
}
\description{
Calculate the mean of a vector
}
\examples{
calc_mean(c(1, 2, 3))       # Returns 2
}
