% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the mean.
}
\description{
Computes the arithmetic mean of a numeric vector.
Missing values (NA) are removed before calculation.
}
\examples{
calc_mean(c(1, 2, 3, 4))
calc_mean(c(1, 2, NA, 4))

}
