% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate arithmetic mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the mean.
}
\description{
Calculates the arithmetic mean of a numeric vector.
Missing values are removed by default.
}
\examples{
calc_mean(c(1, 2, 3)) # Mean is 2
calc_mean(c(1, NA, 3)) # Mean is 2. NA is dropped here
calc_mean(c(5)) # Mean is 5 (single value)
calc_mean(numeric(0)) # Error: Input vector must not be empty.
calc_mean(c(NA, NA, NA)) # Error: Input vector contains only missing values.
}
