% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A scalar numeric value representing the mean.
}
\description{
Computes the arithmetic mean of a numeric vector
}
\examples{
# Basic usage
calc_mean(c(1,2,3))
calc_mean(c(1,2,3,4))

# NA values are removed with explicit information
calc_mean(c(1,2,3, NA))

# All values missing -> error
\dontrun{
calc_mean(c(NA, NA, NA))
}

# Non-numeric input -> error
\dontrun{
calc_mean("a")
}
}
