% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_median}
\alias{calc_median}
\title{Calculate the median}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the median.
}
\description{
Calculates the median of a numeric vector.
}
\examples{
data <- c(1, 2, 3, 4, 5)
calc_median(data)

}
