% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median.R
\name{calc_median}
\alias{calc_median}
\title{Calculate median}
\usage{
calc_median(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector. Must not be empty.}

\item{na.rm}{Logical. Should missing values be removed before calculation?
Defaults to \code{FALSE}.}
}
\value{
A single numeric value representing the median of \code{x}.
}
\description{
Returns median value of a numerical vector, with control over
how \code{NA} values are handled.
}
\examples{
calc_median(c(1, 2, 2, 3, 4, 5, 5, 5, 6, 10))  # 4.5
calc_median(c(1, 2, NA, 4), na.rm = TRUE) # 2

}
\author{
Cara Andrews
}
