% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median.R
\name{calc_median}
\alias{calc_median}
\title{Calculate the Median}
\usage{
calc_median(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric input/vector.}

\item{na.rm}{Handles missing values by setting them to default to FALSE.}
}
\value{
A single numeric value representing the median of the input vector.
}
\description{
Computes the median of a numeric vector.
}
\details{
This function is a wrapper around base R's median function, with additional
input validation for increased robustness.
}
\examples{
calc_median(c(1, 2, 3, 4))
calc_median(c(1, NA, 3), na.rm = TRUE)

}
