% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_median}
\alias{calc_median}
\title{Calculate the median of a vector}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The median of the input vector. Will return error messages if the vector is not numeric or empty
}
\description{
Calculate the median of a vector
}
\examples{
calc_median(c(1, 2, 3))       # Returns 2
}
