% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_median}
\alias{calc_median}
\title{Calculate Median}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the median.
}
\description{
Computes the median of a numeric vector.
Missing values (NA) are removed before calculation.
}
\examples{
calc_median(c(1, 2, 3, 4))
calc_median(c(1, 2, NA, 4))

}
