% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median.R
\name{calc_median}
\alias{calc_median}
\title{Calculate median}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the median.
}
\description{
Calculates the median of a numeric vector.
Missing values are removed by default.
}
\examples{
calc_median(c(1, 2, 3, 4, 5)) # Median is 3
calc_median(c(1, 2, 3, 4)) # Median is 2.5, ie, the average of middle two numbers, 2 and 3.
calc_median(c(1, NA, 3, 4, 5)) # Median is 3.5. NA is dropped
calc_median(c(5)) # Median is 5 (single value)
\dontrun{
calc_median(numeric(0)) # Error: Input vector must not be empty.
calc_median(c(NA, NA, NA)) # Error: Input vector contains only missing values.
}
}
