% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median.R
\name{calc_median}
\alias{calc_median}
\title{Calculate median}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the median.
}
\description{
Calculates the median of a numeric vector.
Missing values are removed by default.
}
\examples{
calc_median(c(1, 2, 3, 4, 5))
calc_median(c(1, 2, 3, 4))
calc_median(c(1, NA, 3, 4, 5))
}
