% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_median}
\alias{calc_median}
\title{Calculate Median}
\usage{
calc_median(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A scalar numeric value representing the median
}
\description{
Computes the median of a numeric vector.
}
\examples{
# Basic usage
calc_median(c(1,2,3))
calc_median(c(1,2,3,4))

# NA values are removed with explicit information
calc_median(c(1,2,3, NA))

# All values missing -> error
\dontrun{
calc_median(c(NA, NA, NA))
}

# Non-numeric input -> error
\dontrun{
calc_median("a")
}
}
