% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate Mode}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Mode(s) of \code{x}. Returns NA if there is no mode.
}
\description{
Calculate Mode
}
