% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate the mode}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value or vector representing the mode.
}
\description{
Calculates the most frequently occurring value(s).
}
\details{
If multiple values have the same highest frequency, all are returned.
If all values occur only once, returns NA.
}
\examples{
calc_mode(c(1, 2, 2, 3))
calc_mode(c(1, 2, 3))

}
