% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate the mode of a vector}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mode of the input vector. Will return error messages if the vector is not numeric, empty, or if there is no mode.
}
\description{
Calculate the mode of a vector
}
\examples{
calc_mode(c(1, 1, 2, 3))      # Returns 1
calc_mode(c(1, 1, 2, 2, 3))   # Returns 1 2
}
