% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate Mode}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the mode.
}
\description{
Computes the mode of a numeric vector.
Missing values (NA) are removed before calculation.
Where Mode is bimodal the first occurring value will be returned
}
\examples{
calc_mode(c(1, 2, 3, 4))
calc_mode(c(1, 2, NA, 4))

}
