% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_q1_and_q3.R
\name{calc_q1}
\alias{calc_q1}
\title{Calculate First Quartile (Q1)}
\usage{
calc_q1(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
The first quartile (Q1).
}
\description{
Calculate First Quartile (Q1)
}
