% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_q1}
\alias{calc_q1}
\title{Calculate the first quartile of a vector}
\usage{
calc_q1(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The first quartile of the input vector. Will return error messages if the vector is not numeric or empty
}
\description{
Calculate the first quartile of a vector
}
\examples{
calc_q1(c(1, 2, 3))       # Returns 1.5
}
