% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_q1}
\alias{calc_q1}
\title{Calculate First Quartile}
\usage{
calc_q1(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the First Quartile.
}
\description{
Computes the First Quartile of a numeric vector.
Missing values (NA) are removed before calculation.
Uses Tukey's Quartile Method
}
\examples{
calc_q1(c(1, 2, 3, 4))
calc_q1(c(1, 2, NA, 4))

}
