% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_q1_and_q3.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate Third Quartile (Q3)}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
The third quartile (Q3).
}
\description{
Calculate Third Quartile (Q3)
}
