% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_1.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate third quartile (Q3)}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value corresponding to the third quartile value.
If x contains missing values, the function will remove them and compute the first quartile for available values.
If x is null or contains a single value, the function will return an error message.
}
\description{
This function calculates the third quartile of a numerical object
}
\examples{
df <- data.frame(
   patid = c("PO1", "P02", "P03", "P04", "P05"),
   glucose = c(5.6, 7.5, 4.3, 8.4, 10.2)
   )

calc_q3(df$glucose)

}
