% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate the third quartile of a vector}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The third quartile of the input vector. Will return error messages if the vector is not numeric or empty
}
\description{
Calculate the third quartile of a vector
}
\examples{
calc_q3(c(1, 2, 3))       # Returns 2.5
}
