% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate Third Quartile}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the Third Quartile.
}
\description{
Computes the Third Quartile of a numeric vector.
Missing values (NA) are removed before calculation.
Uses Tukey's Quartile Method
}
\examples{
calc_q3(c(1, 2, 3, 4))
calc_q3(c(1, 2, NA, 4))

}
