% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_q3.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate third quartile (Q3)}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the third quartile (Q3).
}
\description{
Calculates the third quartile (75th percentile) of a numeric vector.
Missing values are removed by default.
}
\examples{
calc_q3(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
calc_q3(c(1, 2, 2, 3, 4, 5, 5, 5, 6, 10))
calc_q3(c(1, NA, 3, 4, 5))
}
